fix: resolve clippy -D warnings
Satisfies CI: cargo clippy --lib --bins -- -D warnings, and
cargo clippy -p blvm-consensus --lib --all-features -- -D warnings.

Includes profile_log duplicate cfg removal and src/test_helpers adjustments.
Note: cargo clippy --all-targets still reports issues in some integration tests;
CI does not run clippy on tests.

+++ src/bip_validation.rs
@@ -191,7 +191,7 @@ pub fn is_bip54_active(height: Natural, network: crate::types::Network) -> bool 
 #[spec_locked("5.4")]
 pub fn check_bip54_coinbase(coinbase: &Transaction, height: Natural) -> bool {
     let required_lock_time = height.saturating_sub(13);
-    if (coinbase.lock_time as u64) != required_lock_time {
+    if coinbase.lock_time != required_lock_time {
         return false;
     }
     if coinbase.inputs.is_empty() {
